Remove the IJupyterExecutionFactory idea
