+++ src/client/datascience/constants.ts
@@ -140,6 +140,7 @@ export namespace LiveShareJupyterCommands {
     export const isNotebookSupported = 'isNotebookSupported';
     export const isImportSupported = 'isImportSupported';
     export const isKernelCreateSupported = 'isKernelCreateSupported';
+    export const isKernelSpecSupported = 'isKernelSpecSupported';
     export const connectToNotebookServer = 'connectToNotebookServer';
     export const getUsableJupyterPython = 'getUsableJupyterPython';
 }
