+++ src/client/datascience/shiftEnterBanner.ts
@@ -10,7 +10,7 @@ import '../common/extensions';
 import { IConfigurationService, IPersistentStateFactory,
     IPythonExtensionBanner } from '../common/types';
 import * as localize from '../common/utils/localize';
-import { IJupyterExecution, IJupyterExecutionFactory } from './types';
+import { IJupyterExecution } from './types';
 
 export enum InteractiveShiftEnterStateKeys {
     ShowBanner = 'InteractiveShiftEnterBanner'
@@ -32,7 +32,7 @@ export class InteractiveShiftEnterBanner implements IPythonExtensionBanner {
     constructor(
         @inject(IApplicationShell) private appShell: IApplicationShell,
         @inject(IPersistentStateFactory) private persistentState: IPersistentStateFactory,
-        @inject(IJupyterExecutionFactory) private jupyterExecutionFactory: IJupyterExecutionFactory,
+        @inject(IJupyterExecution) private jupyterExecution: IJupyterExecution,
         @inject(IConfigurationService) private configuration: IConfigurationService)
     {
         this.initialize();
